Restore `Parent` constructor for `Local`
This is essentially a revert of 25d86e8e5dd7f5a8b1f7154ca2fa7ed712e0027b

The primary motivation for this is so that paths beginning with `../`
don't get reformatted to begin `./../`.

The secondary motivation for this is to more closely match the standard.
There are fewer special cases in the logic if there is a dedicated
constructor for the `Parent` case.