Add more tests around constructors storing unit
... as requested by @quasicomputational

This exercises both the generic-deriving machinery, which generates
a record of type `{ _1 : {} }` and the `union`/`inputUnion` machinery, which
converts it to an empty alternative

Note that really constructors with one anonymous field should probably
strip the outer `{ _1 : ... }` record, but that is an orthogonal change
which I will postpone until after this upcoming bug-fix release.  However,
if we were to strip them like that then I would expect the generic-deriving
machinery to also generate an empty alternative, just as the
`union`/`inputUnion` machinery.