Downgrade to `yaml-0.10.4.0`
`yaml-0.11.0` split out the `Text.Libyaml` module into a separate `libyaml`
Haskell package.  However, that means that if you depend on the `libyaml`
package then you can no longer use older versions of the `yaml` package
because then they will both conflict due to trying to export `Text.LibYaml`.

That presents one of two options:

* Upgrade `dhall-json` builds for older resolvers to use `yaml-0.11.0`
* Downgrade `dhall-json` builds for newer resolvers to use `yaml-0.10.4.0`

I went with the latter change since this only affects the
`stack.yaml` build and is fairly non-disruptive, whereas upgrading
`stack-lts-6.yaml` to use a newer `yaml` was likely to be more disruptive.

By pinning to an older version of the `yaml` package we no longer need the
`libyaml` dependency in `dhall-json.cabal`, which this change also removes.