Map-like Lambda converts to Map
This allows translating Dhall expressions such as:

```
\(env: < production | development >) -> merge {
	production = "thing1",
	development = "thing2"
} env
```

into the equivalent JSON or YAML map.

These kinds of expressions are a very natural way to encode in Dhall
config files which contain configurations for multiple environments, of
which only one is ever active.  YAML versions of these files tend to be
such a map, since YAML lacks the ability to express it as a function.