Fix LSP diagnostics to trailing whitespace (#974)
* dhall-lsp-server: Revert from Relude to standard Prelude

The code so far used Relude in place of the standard Prelude. Reverting
back to the standard Prelude brings dhall-lsp-server in line with the
rest of the dhall-haskell family, while also reducing the barrier to
contribution.

This change causes a number of additional explicit imports; creating a
"Library" module to gather the common imports could perhaps solve both
problems, i.e. reducing overhead while not having to resort to a custom
prelude. (This is left for a future refactoring.)

*  dhall-lsp-server: Exclude trailing whitespace in diagnostics

Dhall's current parser includes trailing whitespace in source range
annotations. Until now, we simply passed these "loose" ranges on to the
LSP client when generating diagnostic messages; this behaviour is fixed
in the present commit.

Previously (assuming funcTion is misspelled):
 funcTion argument
 ~~~~~~~~~

Now:
  funcTion argument
  ~~~~~~~~

* Refactor lines' to return NonEmpty list

Thanks to Gabriel for pointing this out! Since "lines'" always returns at
least a singleton list containing the empty line we can reflect this in
its type; this allows us to get rid of a test for non-emptiness in
"offsetToPosition".