Fix `Natural/fold` for GHCJS bindings
Fixes https://github.com/dhall-lang/dhall-lang/issues/579

`Natural/fold` was misbehaving on GHCJS due to the use of `Natural`
arithmetic (for some reason).  This is a problem I've seen and fixed
before, but needed to be fixed again after the migration to the new
`Dhall.Eval` normalization engine.

The fix is easy: use `Integer` instead of `Natural` for the accumulator
of the fold and the bug disappears.