Enable Mergify support (#1059)
Related to https://github.com/dhall-lang/dhall-haskell/issues/1058

This is an attempt to test drive the Mergify GitHub app.  Unfortunately, the
only way to know for sure if the app works is to merge this configuration
into `master`, but if it does not work out then I will follow up with a pull
request to revert this change.