Implement optimistic caching (--use-dump)
Normally, we only cache hashed (i.e. `frozen`) imports. This commit adds
a flag `--use-dump` that causes dhall to keep the state of its import
cache around in the file "dhall-cache" in the current directory. The
resulting speedup is the same as would be gained by manually freezing
all imports, with the important difference that whenever a _local_
import changes it and its reverse dependencies are correctly
re-evaluated.

This feature is a prototype! There are a few things left to do:

- Improve the user interface:
  Either separate the flag into `--use-dump` and `--write-dump` to be
  more explicit. In that case using a hidden ".dhall-cache" would be
  reasonable.

  Otherwise use an option `--dump-file = <file>` that takes a file name
  explicitely.

- Pessimistic timestamps:
  Currently we compare the modification times of imports against the
  modification time of the cache dump. Since the dump gets written after
  everything is loaded this creates a race condition. It would be better
  to date back the cache dump to the time before anything is loaded so
  we never miss any changes.

- Code improvements:
  - cacheFromDump is too complicated
  - Possibly move cacheFromDump from Dhall.Main to Dhall.Import?