Fix escaping significant leading whitespace for multiline strings
Fixes https://github.com/dhall-lang/dhall-haskell/issues/1590

This reuses the same logic for computing the longest shared whitespace prefix
that the parser uses in order to ensure consistency of behavior between the
two.  Once we compute the longest shared whitespace prefix we escape *all*
lines with that prefix.  Only one such escape is technically necessary, but
escaping all of them improves horizontal alignment of the result